Implement subclassable ParseState
